viewEntries finished

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -11,19 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('lastname');
-            $table->string('addres');
-            $table->bigInteger('phone');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->string('role');
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('lastname');
+                $table->string('addres');
+                $table->bigInteger('phone');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->integer('role_id')->unsigned();;
+                $table->rememberToken();
+                $table->timestamps();
+
+                //$table->foreign('role_id')->references('id')->on('roles');
+            });
+        }
     }
 
     /**
@@ -31,6 +35,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('users');
+        if (!Schema::hasTable('users')) {
+            Schema::dropIfExists('users');
+        }
     }
 };
